parameterize role for repl, add delay config for repls, better shutdowns

+++ lib/Samaki/Plugin/Repl.rakumod
@@ -3,21 +3,20 @@ use Log::Async;
 use Samaki::Conf;
 use Samaki::Plugin;
 
-unit class Samaki::Plugin::Repl does Samaki::Plugin;
-
-method name { ... }
-method description { ... }
+unit role Samaki::Plugin::Repl[
+  :$name="unnamed",
+  :$cmd=Nil,
+] does Samaki::Plugin;
 
 has Proc::Async $.proc;
 has Promise $.proc-promise;
 has $!pid;
 has $!line-delay-seconds = 1;
 
-has $.command = 'raku';
-
-method write-output {
-  False
-}
+method command { $cmd }
+method write-output { False }
+method name { $name }
+method description { "Run a REPL for $name" }
 
 method start-repl($pane, :$cell) {
   self.stream: [col('info') => "starting repl for {$.name}"];
@@ -28,7 +27,9 @@ method start-repl($pane, :$cell) {
         $!pid = $_;
       }
       whenever $!proc.start(cwd => $cell.data-dir) {
+        info "proc exited for " ~ self.^name;
         $pane.put: "done";
+        $!proc = Nil;
         $!pid = Nil;
         $pane.enable-selection;
       }
@@ -37,8 +38,12 @@ method start-repl($pane, :$cell) {
 }
 
 method execute(Samaki::Cell :$cell, Samaki::Page :$page, Str :$mode, IO::Handle :$out, :$pane, Str :$action) {
+  with $cell.get-conf('delay') -> $delay {
+    $!line-delay-seconds = $delay;
+    self.info: "Seconds between sending lines: $delay";
+  }
   info "launching {$.name} repl";
-  $!proc //= Proc::Async.new: :pty(:rows($pane.height), :cols($pane.width)), $.command;
+  $!proc //= Proc::Async.new: :pty(:rows($pane.height), :cols($pane.width)), self.command;
   unless $!pid {
     self.start-repl($pane, :$cell);
   }
@@ -51,17 +56,23 @@ method execute(Samaki::Cell :$cell, Samaki::Page :$page, Str :$mode, IO::Handle 
 }
 
 method shutdown {
-  .close-stdin with $.proc;
+  with $!proc -> $p {
+    info "close stdin for " ~ self.^name;
+    $!proc.close-stdin;
+  }
   return without $.proc-promise;
   with $.proc-promise {
-    await Promise.anyof($_, Promise.in(2));
+    await Promise.anyof($_, Promise.in(1));
   }
-  if $.proc-promise.status ~~ PromiseStatus::Planned {
+  if $.proc-promise.status ~~ PromiseStatus::Planned  && $!proc {
+    info "sending SIGTERM for " ~ self.^name;
     $!proc.kill(SIGTERM);
     await Promise.anyof($.proc-promise, Promise.in(1));
   }
-  if $.proc-promise.status ~~ PromiseStatus::Planned {
-    $!proc-promise.kill(SIGKILL);
+  if $.proc-promise.status ~~ PromiseStatus::Planned  && $!proc {
+    info "sending SIGKILL for " ~ self.^name;
+    $!proc.kill(SIGKILL);
     await Promise.anyof($.proc-promise, Promise.in(0.5));
   }
+  $!proc = Nil;
 }
